requirements.txt
pandas
duckdb
streamlit
google-generativeai
python-dotenv
langgraph
langchain-google-genai
crewai
